requirements.txt
anthropic>=0.39.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
